refactor: optimize query and query-refresher logic
Removed redundant refresh hook when only data fetching is required.
