+++ .changeset/lovely-camels-tie.md
@@ -1,0 +1,5 @@
+---
+"@reactive-dot/react": patch
+---
+
+Refactored query & query-refresher logic.
