+++ packages/react/src/hooks/use-query.ts
@@ -1,12 +1,10 @@
-import {
-  getQueryInstructionPayloadAtoms,
-  queryPayloadAtomFamily,
-} from "../stores/query.js";
+import { queryPayloadAtomFamily } from "../stores/query.js";
 import type { Falsy, FalsyGuard, FlatHead } from "../types.js";
 import { flatHead, stringify } from "../utils/vanilla.js";
 import type { ChainHookOptions } from "./types.js";
 import { internal_useChainId } from "./use-chain-id.js";
 import { useConfig } from "./use-config.js";
+import { useQueryRefresher } from "./use-query-refresher.js";
 import {
   idle,
   Query,
@@ -19,17 +17,16 @@ import type {
   QueryInstruction,
 } from "@reactive-dot/core/internal.js";
 import { atom, useAtomValue } from "jotai";
-import { useAtomCallback } from "jotai/utils";
-import { useCallback, useMemo } from "react";
+import { useMemo } from "react";
 
 /**
- * Hook for refreshing cached query.
+ * Hook for querying data from chain, and returning the response.
  *
  * @param builder - The function to create the query
  * @param options - Additional options
- * @returns The function to refresh the query
+ * @returns The data response
  */
-export function useQueryRefresher<
+export function useLazyLoadQuery<
   TQuery extends
     | ((
         builder: Query<[], TDescriptor>,
@@ -40,73 +37,6 @@ export function useQueryRefresher<
     : Chains[TChainId],
   TChainId extends ChainId,
 >(builder: TQuery, options?: ChainHookOptions<TChainId>) {
-  const config = useConfig();
-  const chainId = internal_useChainId(options);
-
-  const refresh = useAtomCallback(
-    useCallback(
-      (_, set) => {
-        if (!builder) {
-          return;
-        }
-
-        const query = builder(new Query([]));
-
-        if (!query) {
-          return;
-        }
-
-        const atoms = getQueryInstructionPayloadAtoms(
-          config,
-          chainId,
-          query,
-        ).flat();
-
-        for (const atom of atoms) {
-          if ("write" in atom) {
-            set(atom);
-          }
-        }
-      },
-      [builder, chainId, config],
-    ),
-  );
-
-  return refresh;
-}
-
-/**
- * Hook for querying data from chain, returning the response & a refresher function.
- *
- * @param builder - The function to create the query
- * @param options - Additional options
- * @returns The data response & a function to refresh it
- */
-export function useLazyLoadQueryWithRefresh<
-  TQuery extends
-    | ((
-        builder: Query<[], TDescriptor>,
-      ) => Query<QueryInstruction<TDescriptor>[], TDescriptor> | Falsy)
-    | Falsy,
-  TDescriptor extends TChainId extends void
-    ? CommonDescriptor
-    : Chains[TChainId],
-  TChainId extends ChainId,
->(
-  builder: TQuery,
-  options?: ChainHookOptions<TChainId>,
-): [
-  data: TQuery extends Falsy
-    ? typeof idle
-    : FalsyGuard<
-        ReturnType<Exclude<TQuery, Falsy>>,
-        FlatHead<
-          InferQueryPayload<Exclude<ReturnType<Exclude<TQuery, Falsy>>, Falsy>>
-        >,
-        typeof idle
-      >,
-  refresh: () => void,
-] {
   const config = useConfig();
   const chainId = internal_useChainId(options);
 
@@ -135,29 +65,29 @@ export function useLazyLoadQueryWithRefresh<
     ),
   );
 
-  const data = useMemo(
+  return useMemo(
     () =>
       query && query.instructions.length === 1 ? flatHead(rawData) : rawData,
     [query, rawData],
-  );
-
-  const refresh = useQueryRefresher(builder, options);
-
-  return [
-    // @ts-expect-error complex type
-    data,
-    refresh,
-  ];
+  ) as TQuery extends Falsy
+    ? typeof idle
+    : FalsyGuard<
+        ReturnType<Exclude<TQuery, Falsy>>,
+        FlatHead<
+          InferQueryPayload<Exclude<ReturnType<Exclude<TQuery, Falsy>>, Falsy>>
+        >,
+        typeof idle
+      >;
 }
 
 /**
- * Hook for querying data from chain, and returning the response.
+ * Hook for querying data from chain, returning the response & a refresher function.
  *
  * @param builder - The function to create the query
  * @param options - Additional options
- * @returns The data response
+ * @returns The data response & a function to refresh it
  */
-export function useLazyLoadQuery<
+export function useLazyLoadQueryWithRefresh<
   TQuery extends
     | ((
         builder: Query<[], TDescriptor>,
@@ -167,19 +97,9 @@ export function useLazyLoadQuery<
     ? CommonDescriptor
     : Chains[TChainId],
   TChainId extends ChainId,
->(
-  builder: TQuery,
-  options?: ChainHookOptions<TChainId>,
-): TQuery extends Falsy
-  ? typeof idle
-  : FalsyGuard<
-      ReturnType<Exclude<TQuery, Falsy>>,
-      FlatHead<
-        InferQueryPayload<Exclude<ReturnType<Exclude<TQuery, Falsy>>, Falsy>>
-      >,
-      typeof idle
-    > {
-  const [data] = useLazyLoadQueryWithRefresh(builder, options);
+>(builder: TQuery, options?: ChainHookOptions<TChainId>) {
+  const data = useLazyLoadQuery(builder, options);
+  const refresh = useQueryRefresher(builder, options);
 
-  return data;
+  return [data, refresh] as [data: typeof data, refresh: typeof refresh];
 }
